+++ kamehouse-shell/bin/win/shutdown/suspend.sh
@@ -1,0 +1,14 @@
+#!/bin/bash
+
+# Import kamehouse functions
+source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
+if [ "$?" != "0" ]; then
+  echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing kamehouse-functions.sh\033[0;39m"
+  exit 99
+fi
+
+mainProcess() {
+  cmd.exe "/c rundll32.exe powrprof.dll,SetSuspendState 0,1,0"
+}
+
+main "$@"